# [generator] Explicit initialization of assembly object model