--- a/tools/generator/Java.Interop.Tools.Generator.Importers/Cecil/ManagedInterfaceGen.cs
+++ b/tools/generator/Java.Interop.Tools.Generator.Importers/Cecil/ManagedInterfaceGen.cs
@@ -7,7 +7,7 @@ namespace MonoDroid.Generation
 #if HAVE_CECIL
 	public class ManagedInterfaceGen : InterfaceGen {
 		public ManagedInterfaceGen (TypeDefinition t, CodeGenerationOptions opt)
-			: base (new ManagedGenBaseSupport (t, opt))
+			: base (CecilApiImporter.CreateGenBaseSupport (t, opt))
 		{
 			foreach (var ifaceImpl in t.Interfaces) {
 				AddInterface (ifaceImpl.InterfaceType.FullNameCorrected ());