… tools/generator/Java.Interop.Tools.Generator.Importers/Cecil/ManagedField.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Cecil/ManagedGenBaseSupport.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Cecil/ManagedMethod.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/InterfaceXmlGenBaseSupport.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlClassGen.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlCtor.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlField.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlGenBaseSupport.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlInterfaceGen.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/Xml/XmlMethod.cs
… tools/generator/Java.Interop.Tools.Generator.Importers/XmlApiImporter.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/ClassGen.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/Ctor.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/Field.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/GenBase.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/GenBaseSupport.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/InterfaceGen.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/Method.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/MethodBase.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/Parameter.cs
… tools/generator/Java.Interop.Tools.Generator.ObjectModel/Property.cs
… tools/generator/Java.Interop.Tools.Generator.Transformation/Parser.cs
… tools/generator/Tests/Unit-Tests/ManagedTests.cs
… tools/generator/Tests/Unit-Tests/SupportTypes.cs
… tools/generator/Tests/Unit-Tests/XmlTests.cs
… tools/generator/generator.csproj