As part of the `generator` process we need to build an in-memory model of the assembly we intend to generate sources for.  This model is built from an `api.xml` file for the assembly we intend to generate, and from Cecil for referenced assemblies.  These models look roughly like this:

```
public abstract class ClassGen
{
    public abstract string Visibility { get; }
}

// Cecil
public class ManagedClassGen : ClassGen
{
    TypeDefinition t;
    public override string Visibility => t.Visibility ();
}
```

The issue is Cecil is not thread-safe, and this model means reading any property access is potentially using Cecil, so we cannot safely parallelize any operations.  This PR changes the model to be fully created upfront so that Cecil is no longer used after the model is built, like this:

```
public class ClassGen
{
    public string Visibility { get; set; }
}

public static class CecilApiImporter
{
    public static ClassGen CreateClass (TypeDefinition t)
    {
        return new ClassGen {
            Visibility = t.Visibility ();
        };
    }
}
```

This adds some overhead to the model creation, as it is less lazy now:

Scenario | Original | This PR
------------ | ------------- | -
Mono.Android.dll | 15370ms | 16423ms
Jar2Xml | 5267ms | 5517ms
Class-Parse | 14878ms | 15511ms

This overhead will be mitigated by future PR's that use optimizations this change will enable.

Unfortunately this change was quite invasive and isn't easy to review.  The output is not changed with this PR.  Code changes were verified with existing unit tests and running diffs on `Mono.Android.dll` sources and 20 components from `XamarinComponents` to ensure no output differences.