@jonpryor 2019-07-16
This is partially duplicative of PR #444, which was merged.  This will need to be rebased.